Semester Project
Frequency Division Multiplexing:
Frequency division multiplexing (FDM) is a technique of multiplexing which means 
combining more than one signal over a shared medium. In FDM, signals of different 
frequencies are combined for concurrent transmission.
Concept and Process:
In FDM, the total bandwidth is divided to a set of frequency bands that do not overlap. Each 
of these bands is a carrier of a different signal that is generated and modulated by one of the 
sending devices. The frequency bands are separated from one another by strips of unused 
frequencies called the guard bands, to prevent overlapping of signals.
The modulated signals are combined together using a multiplexer (MUX) in the sending end. 
The combined signal is transmitted over the communication channel, thus allowing multiple 
independent data streams to be transmitted simultaneously. At the receiving end, the 
individual signals are extracted from the combined signal by the process of demultiplexing 
(DEMUX).